[infra] Add native_test_helpers package
