+++ pkgs/native_test_helpers/README.md
@@ -1,0 +1,8 @@
+# Native test helpers
+
+Utils for writing tests in this monorepo. This isn't published, and should only
+be used as a dev dependency for the other packages in this repo, via a path dep.
+
+```shell
+dart pub add "dev:native_test_helpers:{path: ../native_test_helpers}"
+```
