+++ pkgs/native_test_helpers/analysis_options.yaml
@@ -1,0 +1,26 @@
+include: package:dart_flutter_team_lints/analysis_options.yaml
+
+analyzer:
+  errors:
+    todo: ignore
+  language:
+    strict-casts: true
+    strict-inference: true
+    strict-raw-types: true
+  plugins:
+    # - custom_lint # https://github.com/dart-lang/sdk/issues/60784
+
+linter:
+  rules:
+    - avoid_positional_boolean_parameters
+    - dangling_library_doc_comments
+    - directives_ordering
+    - prefer_const_declarations
+    - prefer_expression_function_bodies
+    - prefer_final_in_for_each
+    - prefer_final_locals
+    - use_super_parameters
+
+custom_lint:
+  rules:
+    - avoid_import_outside_src
