+++ pkgs/native_test_helpers/lib/native_test_helpers.dart
@@ -1,0 +1,5 @@
+// Copyright (c) 2025, the Dart project authors. Please see the AUTHORS file
+// for details. All rights reserved. Use of this source code is governed by a
+// BSD-style license that can be found in the LICENSE file.
+
+export 'src/find_package_root.dart';
