+++ pkgs/native_test_helpers/pubspec.yaml
@@ -1,0 +1,9 @@
+name: native_test_helpers
+version: 0.1.0-wip
+repository: https://github.com/dart-lang/native/tree/main/pkgs/native_test_helpers
+description: Helper functions for tests in this monorepo.
+
+publish_to: none
+
+environment:
+  sdk: '>=3.8.0 <4.0.0'
