+++ pkgs/native_toolchain_c/pubspec.yaml
@@ -28,6 +28,8 @@ dev_dependencies:
   collection: ^1.19.1
   custom_lint: ^0.7.5
   dart_flutter_team_lints: ^3.5.1
+  native_test_helpers:
+    path: ../native_test_helpers/
   repo_lint_rules:
     path: ../repo_lint_rules/
   test: ^1.25.15
