+++ pkgs/native_toolchain_c/test/helpers.dart
@@ -8,6 +8,7 @@ import 'dart:io';
 
 import 'package:code_assets/code_assets.dart';
 import 'package:logging/logging.dart';
+import 'package:native_test_helpers/native_test_helpers.dart';
 import 'package:native_toolchain_c/src/native_toolchain/apple_clang.dart';
 import 'package:native_toolchain_c/src/native_toolchain/msvc.dart';
 import 'package:native_toolchain_c/src/utils/run_process.dart';
@@ -80,57 +81,7 @@ Logger _createTestLogger({List<String>? capturedMessages}) =>
         capturedMessages?.add(record.message);
       });
 
-/// Test files are run in a variety of ways, find this package root in all.
-///
-/// Test files can be run from source from any working directory. The Dart SDK
-/// `tools/test.py` runs them from the root of the SDK for example.
-///
-/// Test files can be run from dill from the root of package. `package:test`
-/// does this.
-///
-/// https://github.com/dart-lang/test/issues/110
-Uri findPackageRoot(String packageName) {
-  final script = Platform.script;
-  final fileName = script.name;
-  if (fileName.endsWith('.dart')) {
-    // We're likely running from source in the package somewhere.
-    var directory = script.resolve('.');
-    while (true) {
-      final dirName = directory.name;
-      if (dirName == packageName) {
-        return directory;
-      }
-      final parent = directory.resolve('..');
-      if (parent == directory) break;
-      directory = parent;
-    }
-  } else if (fileName.endsWith('.dill')) {
-    // Probably from the package root.
-    final cwd = Directory.current.uri;
-    final dirName = cwd.name;
-    if (dirName == packageName) {
-      return cwd;
-    }
-  }
-  // Or the workspace root.
-  final cwd = Directory.current.uri;
-  final candidate = cwd.resolve('pkgs/$packageName/');
-  if (Directory.fromUri(candidate).existsSync()) {
-    return candidate;
-  }
-  throw StateError(
-    "Could not find package root for package '$packageName'. "
-    'Tried finding the package root via Platform.script '
-    "'${Platform.script.toFilePath()}' and Directory.current "
-    "'${Directory.current.uri.toFilePath()}'.",
-  );
-}
-
 Uri packageUri = findPackageRoot('native_toolchain_c');
-
-extension on Uri {
-  String get name => pathSegments.where((e) => e != '').last;
-}
 
 /// Archiver provided by the environment.
 ///
