+++ pkgs/objective_c/pubspec.yaml
@@ -31,6 +31,8 @@ dev_dependencies:
   flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
+  native_test_helpers:
+    path: ../native_test_helpers/
   path: ^1.9.0
   test: ^1.21.1
   yaml: ^3.1.0
