+++ pkgs/swift2objc/pubspec.yaml
@@ -26,4 +26,6 @@ environment:
 dev_dependencies:
   args: ^2.6.0
   dart_flutter_team_lints: ^3.5.1
+  native_test_helpers:
+    path: ../native_test_helpers
   test: ^1.21.1
