+++ pkgs/swiftgen/pubspec.yaml
@@ -27,6 +27,8 @@ dependencies:
 
 dev_dependencies:
   dart_flutter_team_lints: ^3.5.1
+  native_test_helpers:
+    path: ../native_test_helpers
   test: ^1.21.1
 
 dependency_overrides:
